test: OpenSearch - reorganize test suite
