+++ integrations/opensearch/tests/test_auth.py
@@ -1,9 +1,14 @@
 from unittest.mock import Mock, patch
 
 import pytest
+from haystack.utils.auth import Secret
 from opensearchpy import AWSV4SignerAsyncAuth, Urllib3AWSV4SignerAuth
 
 from haystack_integrations.document_stores.opensearch.auth import AsyncAWSAuth, AWSAuth
+from haystack_integrations.document_stores.opensearch.document_store import (
+    DEFAULT_MAX_CHUNK_BYTES,
+    OpenSearchDocumentStore,
+)
 
 
 class TestAWSAuth:
@@ -139,3 +144,223 @@ class TestAWSAuth:
         assert async_aws_auth.aws_auth.aws_profile_name.resolve_value() == "some_fake_profile"
         assert async_aws_auth.aws_auth.aws_service == "aoss"
         assert isinstance(async_aws_auth._async_aws_v4_signer_auth, AWSV4SignerAsyncAuth)
+
+
+class TestDocumentStoreWithAuth:
+    @pytest.fixture(autouse=True)
+    def mock_boto3_session(self):
+        with patch("boto3.Session") as mock_client:
+            yield mock_client
+
+    @patch("haystack_integrations.document_stores.opensearch.document_store.OpenSearch")
+    def test_ds_init_with_basic_auth(self, _mock_opensearch_client):
+        document_store = OpenSearchDocumentStore(hosts="testhost", http_auth=("user", "pw"))
+        document_store._ensure_initialized()
+        assert document_store._client
+        _mock_opensearch_client.assert_called_once()
+        assert _mock_opensearch_client.call_args[1]["http_auth"] == ("user", "pw")
+
+    @patch("haystack_integrations.document_stores.opensearch.document_store.OpenSearch")
+    def test_ds_init_without_auth(self, _mock_opensearch_client):
+        document_store = OpenSearchDocumentStore(hosts="testhost")
+        document_store._ensure_initialized()
+        assert document_store._client
+        _mock_opensearch_client.assert_called_once()
+        assert _mock_opensearch_client.call_args[1]["http_auth"] is None
+
+    @patch("haystack_integrations.document_stores.opensearch.document_store.OpenSearch")
+    def test_ds_init_aws_auth(self, _mock_opensearch_client):
+        document_store = OpenSearchDocumentStore(
+            hosts="testhost",
+            http_auth=AWSAuth(aws_region_name=Secret.from_token("dummy-region")),
+            use_ssl=True,
+            verify_certs=True,
+        )
+        document_store._ensure_initialized()
+        assert document_store._client
+        _mock_opensearch_client.assert_called_once()
+        assert isinstance(_mock_opensearch_client.call_args[1]["http_auth"], AWSAuth)
+        assert _mock_opensearch_client.call_args[1]["use_ssl"] is True
+        assert _mock_opensearch_client.call_args[1]["verify_certs"] is True
+
+    @patch("haystack_integrations.document_stores.opensearch.document_store.OpenSearch")
+    def test_ds_from_dict_basic_auth(self, _mock_opensearch_client):
+        document_store = OpenSearchDocumentStore.from_dict(
+            {
+                "type": "haystack_integrations.document_stores.opensearch.document_store.OpenSearchDocumentStore",
+                "init_parameters": {
+                    "hosts": "testhost",
+                    "http_auth": ["user", "pw"],
+                    "use_ssl": True,
+                    "verify_certs": True,
+                },
+            }
+        )
+        document_store._ensure_initialized()
+        assert document_store._client
+        _mock_opensearch_client.assert_called_once()
+        assert _mock_opensearch_client.call_args[1]["http_auth"] == ["user", "pw"]
+
+    @patch("haystack_integrations.document_stores.opensearch.document_store.OpenSearch")
+    def test_ds_from_dict_aws_auth(self, _mock_opensearch_client, monkeypatch: pytest.MonkeyPatch):
+        monkeypatch.setenv("AWS_DEFAULT_REGION", "dummy-region")
+        document_store = OpenSearchDocumentStore.from_dict(
+            {
+                "type": "haystack_integrations.document_stores.opensearch.document_store.OpenSearchDocumentStore",
+                "init_parameters": {
+                    "hosts": "testhost",
+                    "http_auth": {
+                        "type": "haystack_integrations.document_stores.opensearch.auth.AWSAuth",
+                        "init_parameters": {},
+                    },
+                    "use_ssl": True,
+                    "verify_certs": True,
+                },
+            }
+        )
+        document_store._ensure_initialized()
+        assert document_store._client
+        _mock_opensearch_client.assert_called_once()
+        assert isinstance(_mock_opensearch_client.call_args[1]["http_auth"], AWSAuth)
+        assert _mock_opensearch_client.call_args[1]["use_ssl"] is True
+        assert _mock_opensearch_client.call_args[1]["verify_certs"] is True
+
+    @patch("haystack_integrations.document_stores.opensearch.document_store.OpenSearch")
+    def test_ds_to_dict_basic_auth(self, _mock_opensearch_client):
+        document_store = OpenSearchDocumentStore(hosts="some hosts", http_auth=("user", "pw"))
+        res = document_store.to_dict()
+        assert res == {
+            "type": "haystack_integrations.document_stores.opensearch.document_store.OpenSearchDocumentStore",
+            "init_parameters": {
+                "embedding_dim": 768,
+                "hosts": "some hosts",
+                "index": "default",
+                "mappings": {
+                    "dynamic_templates": [
+                        {"strings": {"mapping": {"type": "keyword"}, "match_mapping_type": "string"}}
+                    ],
+                    "properties": {
+                        "content": {"type": "text"},
+                        "embedding": {"dimension": 768, "index": True, "type": "knn_vector"},
+                    },
+                },
+                "max_chunk_bytes": DEFAULT_MAX_CHUNK_BYTES,
+                "method": None,
+                "settings": {"index.knn": True},
+                "return_embedding": False,
+                "create_index": True,
+                "http_auth": ("user", "pw"),
+                "use_ssl": None,
+                "verify_certs": None,
+                "timeout": None,
+            },
+        }
+
+    @patch("haystack_integrations.document_stores.opensearch.document_store.OpenSearch")
+    def test_ds_to_dict_aws_auth(self, _mock_opensearch_client, monkeypatch: pytest.MonkeyPatch):
+        monkeypatch.setenv("AWS_DEFAULT_REGION", "dummy-region")
+        document_store = OpenSearchDocumentStore(hosts="some hosts", http_auth=AWSAuth())
+        res = document_store.to_dict()
+        assert res == {
+            "type": "haystack_integrations.document_stores.opensearch.document_store.OpenSearchDocumentStore",
+            "init_parameters": {
+                "embedding_dim": 768,
+                "hosts": "some hosts",
+                "index": "default",
+                "mappings": {
+                    "dynamic_templates": [
+                        {"strings": {"mapping": {"type": "keyword"}, "match_mapping_type": "string"}}
+                    ],
+                    "properties": {
+                        "content": {"type": "text"},
+                        "embedding": {"dimension": 768, "index": True, "type": "knn_vector"},
+                    },
+                },
+                "max_chunk_bytes": DEFAULT_MAX_CHUNK_BYTES,
+                "method": None,
+                "settings": {"index.knn": True},
+                "return_embedding": False,
+                "create_index": True,
+                "http_auth": {
+                    "type": "haystack_integrations.document_stores.opensearch.auth.AWSAuth",
+                    "init_parameters": {
+                        "aws_access_key_id": {"type": "env_var", "env_vars": ["AWS_ACCESS_KEY_ID"], "strict": False},
+                        "aws_secret_access_key": {
+                            "type": "env_var",
+                            "env_vars": ["AWS_SECRET_ACCESS_KEY"],
+                            "strict": False,
+                        },
+                        "aws_session_token": {"type": "env_var", "env_vars": ["AWS_SESSION_TOKEN"], "strict": False},
+                        "aws_region_name": {"type": "env_var", "env_vars": ["AWS_DEFAULT_REGION"], "strict": False},
+                        "aws_profile_name": {"type": "env_var", "env_vars": ["AWS_PROFILE"], "strict": False},
+                        "aws_service": "es",
+                    },
+                },
+                "use_ssl": None,
+                "verify_certs": None,
+                "timeout": None,
+            },
+        }
+
+    @patch("haystack_integrations.document_stores.opensearch.document_store.OpenSearch")
+    def test_ds_init_with_env_var_secrets(self, _mock_opensearch_client, monkeypatch):
+        """Test the default initialization using environment variables"""
+        monkeypatch.setenv("OPENSEARCH_USERNAME", "user")
+        monkeypatch.setenv("OPENSEARCH_PASSWORD", "pass")
+
+        document_store = OpenSearchDocumentStore(hosts="testhost")
+        document_store._ensure_initialized()
+        assert document_store._client
+        _mock_opensearch_client.assert_called_once()
+        assert _mock_opensearch_client.call_args[1]["http_auth"] == ["user", "pass"]
+
+    @patch("haystack_integrations.document_stores.opensearch.document_store.OpenSearch")
+    def test_ds_init_with_missing_env_vars(self, _mock_opensearch_client):
+        """Test that auth is None when environment variables are missing"""
+        document_store = OpenSearchDocumentStore(hosts="testhost")
+        document_store._ensure_initialized()
+        assert document_store._client
+        _mock_opensearch_client.assert_called_once()
+        assert _mock_opensearch_client.call_args[1]["http_auth"] is None
+
+    @patch("haystack_integrations.document_stores.opensearch.document_store.OpenSearch")
+    def test_ds_to_dict_with_env_var_secrets(self, _mock_opensearch_client, monkeypatch):
+        """Test serialization with environment variables"""
+        monkeypatch.setenv("OPENSEARCH_USERNAME", "user")
+        monkeypatch.setenv("OPENSEARCH_PASSWORD", "pass")
+
+        document_store = OpenSearchDocumentStore(hosts="testhost")
+        serialized = document_store.to_dict()
+
+        assert "http_auth" in serialized["init_parameters"]
+        auth = serialized["init_parameters"]["http_auth"]
+        assert isinstance(auth, list)
+        assert len(auth) == 2
+        # Check that we have two Secret dictionaries with correct env vars
+        assert auth[0]["type"] == "env_var"
+        assert auth[0]["env_vars"] == ["OPENSEARCH_USERNAME"]
+        assert auth[1]["type"] == "env_var"
+        assert auth[1]["env_vars"] == ["OPENSEARCH_PASSWORD"]
+
+    @patch("haystack_integrations.document_stores.opensearch.document_store.OpenSearch")
+    def test_ds_from_dict_with_env_var_secrets(self, _mock_opensearch_client, monkeypatch):
+        """Test deserialization with environment variables"""
+        # Set environment variables so the secrets resolve properly
+        monkeypatch.setenv("OPENSEARCH_USERNAME", "user")
+        monkeypatch.setenv("OPENSEARCH_PASSWORD", "pass")
+
+        data = {
+            "type": "haystack_integrations.document_stores.opensearch.document_store.OpenSearchDocumentStore",
+            "init_parameters": {
+                "hosts": "testhost",
+                "http_auth": [
+                    {"type": "env_var", "env_vars": ["OPENSEARCH_USERNAME"], "strict": False},
+                    {"type": "env_var", "env_vars": ["OPENSEARCH_PASSWORD"], "strict": False},
+                ],
+            },
+        }
+        document_store = OpenSearchDocumentStore.from_dict(data)
+        document_store._ensure_initialized()
+        assert document_store._client
+        _mock_opensearch_client.assert_called_once()
+        assert _mock_opensearch_client.call_args[1]["http_auth"] == ["user", "pass"]
